CRA-180: Fixed handling of missingvalues in CRAVA when using STORM- and sgri-files. Now use IsMissing in stormcontgrid class.
model.cpp: Removed wrong SetMissingCode in readStormFile.
fftgrid.cpp: Changed missingtest in fillInFromStorm.

crava.cpp, blockedlogs.cpp - esthetic fixes.

Changed output of return value from stormcontgrid to float, hence the updated synthetic seismic files.

git-svn-id: https://code.nr.no/scm/svn/crava/trunk@855 176de4d3-4747-0410-aa96-47f6d6634214

+++ src/blockedlogs.cpp
@@ -1744,7 +1744,7 @@ void BlockedLogs::findSeismicGradient(FFTGrid                  ** seisCube,
         seisTrace = seisCube[l]->getRealTrace2(i0+i, j0+j);
         smoothTrace(seisTrace);
         if(i==0){
-          for(int s = 0; s < seisTrace.size(); s++)
+          for(size_t s = 0; s < seisTrace.size(); s++)
             out << seisTrace[s] << std::endl;       
         }
